touchup: extract url param hooks from nav store
