+++ src/pages/playground.page.tsx
@@ -6,8 +6,8 @@ import { useEffect, useState } from "react";
 import { Loading } from "../web/common/components/Loading/Loading";
 import { populateFromLocalStorage } from "../web/topic/store/loadActions";
 import { playgroundUsername } from "../web/topic/store/store";
-import { useSyncSearchParamsWithStore } from "../web/view/navigateStore";
 import { setInitialPerspective } from "../web/view/perspectiveStore";
+import { useSyncSearchParamsWithStore } from "../web/view/urlParamHooks";
 
 // Don't render the workspace server-side.
 // Known reasons:
